file keranjang

+++ function/penjual/penjual.h
@@ -5,7 +5,7 @@ typedef struct {
     int id;
     char name[50];
     char category[30];
-    float price;
+    int price;
     int stock;
     int id_penjual;
 } Product;
